Renamed files to follow snake_case
