+++ lib/models/search_model.dart
