+++ lib/pages/searchPage/category_filter_chips.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
-import 'package:nestify/models/searchModel.dart';
+import 'package:nestify/models/search_model.dart';
 
 class CategoryFilterChips extends StatefulWidget {
   const CategoryFilterChips({super.key});
