+++ lib/pages/searchPage/filter_modal_bottomsheet.dart
@@ -1,8 +1,8 @@
 import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
-import 'package:nestify/models/searchModel.dart';
-import 'package:nestify/pages/searchPage/categoryFilterChips.dart';
-import 'package:nestify/pages/searchPage/materialFilterChips.dart';
+import 'package:nestify/models/search_model.dart';
+import 'package:nestify/pages/searchPage/category_filter_chips.dart';
+import 'package:nestify/pages/searchPage/material_filter_chips.dart';
 import 'package:go_router/go_router.dart';
 
 void filterModalBottomsheet(BuildContext context) {
